polyforms tests
